requirements.txt
numpy
torch
scikit-learn
tqdm
Pillow